bug: useSearchParams() Suspense 이슈로 빌드가 안돼서 서버컴포넌트에서 searchParams 받는 방식으로 변경하였음
